+++ st2actions/st2actions/runners/httprunner.py
@@ -20,7 +20,6 @@ import uuid
 
 import requests
 from oslo_config import cfg
-from six.moves.urllib import parse as urlparse
 
 from st2actions.runners import ActionRunner
 from st2common import __version__ as st2_version
@@ -71,7 +70,6 @@ class HttpRunner(ActionRunner):
                                                           self._on_behalf_user)
         self._url = self.runner_parameters.get(RUNNER_URL, None)
         self._headers = self.runner_parameters.get(RUNNER_HEADERS, {})
-        self._headers = self._params_to_dict(self._headers)
 
         self._cookies = self.runner_parameters.get(RUNNER_COOKIES, None)
         self._allow_redirects = self.runner_parameters.get(RUNNER_ALLOW_REDIRECTS, False)
@@ -90,7 +88,6 @@ class HttpRunner(ActionRunner):
         timeout = float(action_parameters.get(ACTION_TIMEOUT, self._timeout))
         method = action_parameters.get(ACTION_METHOD, None)
         params = action_parameters.get(ACTION_QUERY_PARAMS, None)
-        params = self._params_to_dict(params)
         auth = action_parameters.get(ACTION_AUTH, {})
 
         file_name = action_parameters.get(FILE_NAME, None)
@@ -126,15 +123,6 @@ class HttpRunner(ActionRunner):
                           headers=headers, cookies=self._cookies, auth=auth,
                           timeout=timeout, allow_redirects=self._allow_redirects,
                           proxies=proxies, files=files, verify=self._verify_ssl_cert)
-
-    def _params_to_dict(self, params):
-        if not params:
-            return {}
-
-        if isinstance(params, dict):
-            return params
-
-        return dict(urlparse.parse_qsl(params, keep_blank_values=True, strict_parsing=True))
 
     @staticmethod
     def _get_result_status(status_code):